Database/InstaLeadMagnitBot/requirements.txt
# Instagram API
instagrapi>=2.0.0

# Telegram Bot
python-telegram-bot>=20.0

# Database
SQLAlchemy>=2.0.0
aiosqlite>=0.19.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Logging
loguru>=0.7.0

# Google Sheets
gspread>=5.0.0
google-auth>=2.0.0
